+++ src/app/providers/route/app-layout.tsx
@@ -4,8 +4,8 @@ import useMediaQuery from '@mui/material/useMediaQuery';
 import Box from '@mui/material/Box';
 import Toolbar from '@mui/material/Toolbar';
 import { Outlet } from 'react-router';
-import Header from '../../../widgets/header/header';
-import Sidebar from '../../../widgets/sidebar/sidebar';
+import { Header, Sidebar } from '@widgets/';
+
 
 export default function DashboardLayout() {
   const theme = useTheme();
