+++ src/app/providers/route/index.ts
@@ -1,0 +1,1 @@
+export { router } from './router';
