+++ src/app/providers/store/index.ts
@@ -1,0 +1,1 @@
+export { store, useAppDispatch, useAppSelector } from './store';
