+++ src/features/auth/component/index.ts
@@ -1,0 +1,1 @@
+export { LoginButton } from "./login-button/login-button";
