+++ src/features/auth/index.ts
@@ -1,0 +1,2 @@
+export * from './component';
+export * from './model';
