+++ src/features/auth/model/index.ts
@@ -1,0 +1,2 @@
+export { authSlice, logout } from "./auth-slice";
+export { handleCheckAuth, handleLogin } from "./thunks";
