+++ src/features/light-macros/components/cinema-multi-select-buttons.ts/index.ts
@@ -1,0 +1,1 @@
+export { CinemaMultiSelectButtons } from './cinema-multi-select-buttons';
