+++ src/features/light-macros/components/cinema-styled-button/index.ts
@@ -1,0 +1,1 @@
+export { CinemaStyledButton } from './cinema-styled-button';
