+++ src/features/light-macros/components/task-light-macros/index.ts
@@ -1,0 +1,1 @@
+export { TaskLightMacros } from './task-light-macros';
