+++ src/main.tsx
@@ -1,7 +1,7 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import App from './app/App.tsx'
 import { StyledEngineProvider } from '@mui/material/styles'
+import { App } from '@app/'
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
