+++ src/pages/index.ts
@@ -1,0 +1,3 @@
+export * from './info-page';
+export * from './light-macros-page';
+export * from './settings-page';
