+++ src/pages/info-page/index.ts
@@ -1,0 +1,1 @@
+export { InfoPage } from './info-page';
