+++ src/pages/light-macros-page/index.ts
@@ -1,0 +1,1 @@
+export { LightMacrosPage } from "./light-macros-page";
