+++ src/pages/settings-page/index.ts
@@ -1,0 +1,1 @@
+export {SettingsPage} from './settings-page';
