+++ src/widgets/index.ts
@@ -1,0 +1,3 @@
+export * from './header';
+export * from './sidebar';
+export * from './task-list';
