+++ src/widgets/task-list/index.ts
@@ -1,0 +1,1 @@
+export { TaskList } from './task-list';
