+++ vite.config.ts
@@ -1,7 +1,12 @@
 import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
+import tsconfigPaths from "vite-tsconfig-paths";
 
-// https://vite.dev/config/
+
 export default defineConfig({
-  plugins: [react()],
+  plugins: [
+		react(),
+		tsconfigPaths()
+	],
 })
+
